Moved testing of conditional_spherical_isolation into its own dedicated testing module
